Minor type correction

+++ alibi_detect/cd/fet.py
@@ -184,7 +184,7 @@ def fisher_exact(a: np.ndarray, b: np.ndarray, c: np.ndarray, d: np.ndarray, alt
 
 
 @nb.njit(cache=True)
-def _betaln(a: float, b: float) -> float:
+def _betaln(a: int, b: int) -> float:
     """
     Natural logarithm of absolute value of beta function. Equivalent to scipy.special.betaln, but numba accelerated.
     This is to avoid using the numba-scipy package which is currently a little slow to adopt new scipy versions.
@@ -201,21 +201,24 @@ def _betaln(a: float, b: float) -> float:
     -------
     The result of ln(abs(beta(a,b))).
     """
+    # Special case: -ve values
     if a <= 0:
         if 1-a-b > 0:
             return _betaln(1-a-b, b)
     if b <= 0:
         if 1-a-b > 0:
             return _betaln(1-a-b, a)
+    # Special case: a < b
     if a < b:
         a, b = b, a
 
-    # asymptotic expansion for |a| >> |b|
+    # Special case: asymptotic expansion for |a| >> |b|
     # TODO - Performing an asymptotic expansion via Stirling's series is more precise
     #  for large a (or b). But current implementation appears to be sufficient. Can improve in future.
     # if abs(a) > 1e5*abs(b) and abs(a) > 1e5:
     #    return _loggammadiv(b, a) + math.lgamma(b)
 
+    # Final evaluation
     ya = math.lgamma(abs(a))
     yb = math.lgamma(abs(b))
     yab = math.lgamma(a+b)
